Use geolocation for determining wards and precincts
Instead of relying on the wards and precincts encoded in the Street
Address Mapping CSV, which have tons and tons of errors in them, use
the official ward and precinct boundaries and the latitude / longitude
coordinates of the city's addresses to determine address wards and
precincts. This is somewhat slower but _much_ more accurate, perhaps
even accurate enough for production use, though there are still some
inconsistencies in the SAM data.